--- MovieQuiz/Models/QuizQuestion.swift
@@ -1,14 +1,0 @@
-//
-//  QuizQuestion.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 17.01.2024.
-//
-
-import Foundation
-
-struct QuizQuestion {
-  let image: String
-  let text: String
-  let correctAnswer: Bool
-}
