--- MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  QuizResultsViewModel.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 17.01.2024.
-//
-
-import Foundation
-
-struct QuizResultsViewModel {
-  let title: String
-  let text: String
-  let buttonText: String
-}
