--- MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  QuizStepViewModel.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 17.01.2024.
-//
-import Foundation
-import UIKit
-
-struct QuizStepViewModel {
-  let image: UIImage
-  let question: String
-  let questionNumber: String
-}
