--- MovieQuiz/Services/AlertProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  AlertProtocol.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 29.02.2024.
-//
-
-import Foundation
-
-protocol AlertProtocol {
-    func show(alertModel: AlertModel)
-}
