--- MovieQuiz/Services/QuestionFactoryDelegate.swift
@@ -1,11 +1,0 @@
-//
-//  QuestionFactoryDelegate.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 28.02.2024.
-//
-
-import Foundation
-protocol QuestionFactoryDelegate: AnyObject {
-    func didReceiveNextQuestion(question: QuizQuestion?)
-}
