--- MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  QuestionFactoryProtocol.swift
-//  MovieQuiz
-//
-//  Created by Yerman Ibragimuly on 27.02.2024.
-//
-
-import Foundation
-
-protocol QuestionFactoryProtocol {
-    func requestNextQuestion()
-}
